requirements.txt
pyyaml==6.0
numpy==1.22.3
torch==1.11.0
transformers==4.19.0
tqdm==4.64.0
nltk==3.7